Basically undone these commits:
82da2651b2034114fe6305b922471c9993823092
780d221b2f3a8a112fdcce888d3f7e515b805b56

Instead of wait()-ing which has just made the problem worse, the
socket used for passing data between the threads needs to be closed.

The second is yet to be solved, onPause() is called even if the
Activity is being canceled, need to check somehow why this is needed
for saving the instance state.

Signed-off-by: Balint Kovacs <blint@blint.hu>